+++ src/main/java/com/github/jeffmagina/restaurant/cashier/Cashier.java
@@ -1,79 +1,95 @@
 package com.github.jeffmagina.restaurant.cashier;
 
+import java.io.File;
+import java.util.ArrayList;
+
+import com.github.jeffmagina.io.IO;
 import com.github.jeffmagina.restaurant.customerorder.CustomerOrder;
 import com.github.jeffmagina.restaurant.menu.Menu;
 //import com.github.jeffmagina.restaurant.orderhistory.OrderHistory;
 
 public class Cashier {
 
-    private String name;
+	private String name;
+	
+	private ArrayList<CustomerOrder> custOrders = new ArrayList<>();
+	public Menu menu = new Menu();
+	private CustomerOrder custOrder = new CustomerOrder();
 
-    public Cashier(String name){
-        this.name = name;
-    }
+	public Cashier(String name) {
+		this.name = name;
+	}
+	
+	public void interaction() {
+		File orderForm = new File("OrderForm.txt");
+		
+		IO parser = new IO();
+		custOrder = parser.ParseCustOrderForm(orderForm);
+		
+		greeting();
+		takeOrder();
+		transaction();
+		storeOrder();
+		
+	}
 
-    public void greeting(CustomerOrder custOrder, Menu menu){
-        System.out.println("\nHello " + custOrder.name 
-        + "! My name is " + this.name + "!"
-        + "\nWelcome to Jeff's Restaurant! \n" 
-        + "Our menu options are: \n"); 
-        menu.displayMenu();
-        
-    }
+	private void greeting() {	
+		System.out.println("\nHello " + custOrder.name + "! My name is " + this.name + "!"
+				+ "\nWelcome to Jeff's Restaurant! \n" + "Our menu options are: \n");
+		this.menu.displayMenu();
 
-    public void takeOrder(CustomerOrder custOrder){
-        System.out.println("\nThe order you have placed is: ");
-        custOrder.displayCustOrder();
-    }
+	}
 
-    public void transaction(CustomerOrder custOrder, Menu menu){
-        custOrder.orderCost = calcCost(custOrder, menu);
-        System.out.println("\nThe cost of your order is: " + custOrder.orderCost);
-        System.out.println("\nYour payment amount was: " + custOrder.paymentAmount);
+	private void takeOrder() {	
+		System.out.println("\nThe order you have placed is: ");
+		custOrder.displayCustOrder();
+	}
 
-        custOrder.changeGiven = custOrder.paymentAmount - custOrder.orderCost;
+	private void transaction() {
+		custOrder.orderCost = calcCost(custOrder);
+		System.out.println("\nThe cost of your order is: " + custOrder.orderCost);
+		System.out.println("\nYour payment amount was: " + custOrder.paymentAmount);
 
-        if(custOrder.changeGiven >= 0){
-            System.out.println("\nYour change is: " + custOrder.changeGiven);
-        } else{
-            System.out.println("\nInsufficient funds! Come back sufficient funds!\n");
-            System.exit(0);
-        }
+		custOrder.changeGiven = custOrder.paymentAmount - custOrder.orderCost;
 
-    }
+		if (custOrder.changeGiven >= 0) {
+			System.out.println("\nYour change is: " + custOrder.changeGiven);
+		} else {
+			System.out.println("\nInsufficient funds! Come back sufficient funds!\n");
+			System.exit(0);
+		}
 
-    private double calcCost(CustomerOrder order,Menu menu){
-        double cost = 0.0;
-            for(int i = 0; i<order.size(); i++){
-                for(int j = 0; j<menu.size(); j++){
-                    if(order.getItem(i).equals(menu.getItem(j).name)){                   
-                        cost = cost + menu.getItem(j).cost;
-                    }
-                } 
-            }
-            return cost;
-    }
+	}
 
-    public void storeOrder(CustomerOrder customerOrder){
-
-        //Maybe make an order class with this info
+	private void storeOrder(){
         System.out.println("\nOrder Stored: " );
         System.out.print("Customer Name: ");
-        System.out.println(customerOrder.name);
+        System.out.println(custOrder.name);
         System.out.print("Customer Order: ");
-        System.out.println(customerOrder.order);
+        System.out.println(custOrder.order);
         System.out.print("Order Cost: ");
-        System.out.println(customerOrder.orderCost);
+        System.out.println(custOrder.orderCost);
         System.out.print("Customer Payment: ");
-        System.out.println(customerOrder.paymentAmount);
+        System.out.println(custOrder.paymentAmount);
         System.out.print("Customer Change: ");
-        System.out.println(customerOrder.changeGiven);
-    }
-
-    public void displayOrderHistory(){
-       // OrderHistory orderHistory = new OrderHistory();
+        System.out.println(custOrder.changeGiven);
         
-        //store in database
+		File history = new File("history.txt");
+    	IO writer= new IO();
+    	
+        writer.write(history, custOrder);
+        writer.read(history, custOrder, this.custOrders);
     }
-
+	
+	private double calcCost(CustomerOrder order) {
+		double cost = 0.0;
+		for (int i = 0; i < order.size(); i++) {
+			for (int j = 0; j < this.menu.size(); j++) {
+				if (order.getItem(i).equals(this.menu.getItem(j).name)) {
+					cost = cost + this.menu.getItem(j).cost;
+				}
+			}
+		}
+		return cost;
+	}
 }